refactoring, fixed Release build, adapted scripts to new syntax
