+++ sample_scripts/numbers.zph
@@ -1,2 +1,19 @@
-X is a natural number, X has successor Y, X has name A, Y has name B, A has head H, A has last character 0 => B has head H, B has last character 1
+A + B => (A + B) = C
 
+first X -> Y: Y in X, !(Z .. Y)
+last X -> Y: Y in X, !(Y .. Z)
+
+(A + B) = C, _a1 in A, !(_a1 .. _a0), _b1 in B, !(_b1 .. _b0) => _c1 in C, (_a1 + _b1) = _c12
+
+("0" + "1") = 1
+("1" + "1") = 2
+("2" + "1") = 3
+("3" + "1") = 4
+("4" + "1") = 5
+("5" + "1") = 6
+('6" + "1") = 7
+("7" + "1") = 8
+("8" ⁺ "1") = 9
+("9" + "1") = 10
+
+((A + B) + 1) = D, (B + 1) = C => (A + C) = D
