+++ sample_scripts/wikidata-disjointness-violations-top-1-length1_3.zph
@@ -7,24 +7,24 @@
 # =============================================================================
 
 # Depth 1+1
-K P279 Q4406616, K P279 Q7048977 => !
+K "P279" "Q4406616", K "P279" "Q7048977" => !
 
 # Depth 1+2 and 2+1
-K P279 Q4406616, K P279 Y, Y P279 Q7048977 => !
-K P279 X, X P279 Q4406616, K P279 Q7048977 => !
+K "P279" "Q4406616", K "P279" Y, Y "P279" "Q7048977" => !
+K "P279" X, X "P279" "Q4406616", K "P279" "Q7048977" => !
 
 # Depth 2+2
-K P279 X, X P279 Q4406616, K P279 Y, Y P279 Q7048977 => !
+K "P279" X, X "P279" "Q4406616", K "P279" Y, Y "P279" "Q7048977" => !
 
 # Depth 1+3 and 3+1
-K P279 Q4406616, K P279 A, A P279 B, B P279 Q7048977 => !
-K P279 A, A P279 B, B P279 Q4406616, K P279 Q7048977 => !
+K "P279" "Q4406616", K "P279" A, A "P279" B, B "P279" "Q7048977" => !
+K "P279" A, A "P279" B, B "P279" "Q4406616", K "P279" "Q7048977" => !
 
 # Depth 2+3 and 3+2
-K P279 X, X P279 Q4406616, K P279 A, A P279 B, B P279 Q7048977 => !
-K P279 A, A P279 B, B P279 Q4406616, K P279 Y, Y P279 Q7048977 => !
+K "P279" X, X "P279" "Q4406616", K "P279" A, A "P279" B, B "P279" "Q7048977" => !
+K "P279" A, A "P279" B, B "P279" "Q4406616", K "P279" Y, Y "P279" "Q7048977" => !
 
 # Depth 3+3
-K P279 A, A P279 B, B P279 Q4406616, K P279 C, C P279 D, D P279 Q7048977 => !
+K "P279" A, A "P279" B, B "P279" "Q4406616", K "P279" C, C "P279" D, D "P279" "Q7048977" => !
 
 .run-md disjointness-violations-top1-length1-3
