+++ sample_scripts/wikidata-disjointness-violations-top-4-length1_2.zph
@@ -2,27 +2,27 @@
 .lang wikidata
 
 # Top 1: concrete object vs abstract entity (93% of the total culprits)
-K P279 Q4406616, K P279 Y, Y P279 Q7048977 => !
-K P279 Q7048977, K P279 Y, Y P279 Q4406616 => !
-X P31 Q4406616, Y P31 Q7048977, X P31 Y => !
-X P31 Q7048977, Y P31 Q4406616, X P31 Y => !
+K "P279" "Q4406616", K "P279" Y, Y "P279" "Q7048977" => !
+K "P279" "Q7048977", K "P279" Y, Y "P279" "Q4406616" => !
+X "P31" "Q4406616", Y "P31" "Q7048977", X "P31" Y => !
+X "P31" "Q7048977", Y "P31" "Q4406616", X "P31" Y => !
 
 # Top 2: artificial object vs natural object
-K P279 Q16686448, K P279 Y, Y P279 Q1970309 => !
-K P279 Q1970309, K P279 Y, Y P279 Q16686448 => !
-X P31 Q16686448, X P31 Y, Y P31 Q1970309 => !
-X P31 Q1970309, X P31 Y, Y P31 Q16686448 => !
+K "P279" "Q16686448", K "P279" Y, Y "P279" "Q1970309" => !
+K "P279" "Q1970309", K "P279" Y, Y "P279" "Q16686448" => !
+X "P31" "Q16686448", X "P31" Y, Y "P31" "Q1970309" => !
+X "P31" "Q1970309", X "P31" Y, Y "P31" "Q16686448" => !
 
 # Top 3: natural physical object vs artificial physical object
-K P279 Q16686022, K P279 Y, Y P279 Q8205328 => !
-K P279 Q8205328, K P279 Y, Y P279 Q16686022 => !
-X P31 Q16686022, X P31 Y, Y P31 Q8205328 => !
-X P31 Q8205328, X P31 Y, Y P31 Q16686022 => !
+K "P279" "Q16686022", K "P279" Y, Y "P279" "Q8205328" => !
+K "P279" "Q8205328", K "P279" Y, Y "P279" "Q16686022" => !
+X "P31" "Q16686022", X "P31" Y, Y "P31" "Q8205328" => !
+X "P31" "Q8205328", X "P31" Y, Y "P31" "Q16686022" => !
 
 # Top 4: person vs organization
-K P279 Q215627, K P279 Y, Y P279 Q43229 => !
-K P279 Q43229, K P279 Y, Y P279 Q215627 => !
-X P31 Q215627, X P31 Y, Y P31 Q43229 => !
-X P31 Q43229, X P31 Y, Y P31 Q215627 => !
+K "P279" "Q215627", K "P279" Y, Y "P279" "Q43229" => !
+K "P279" "Q43229", K "P279" Y, Y "P279" "Q215627" => !
+X "P31" "Q215627", X "P31" Y, Y "P31" "Q43229" => !
+X "P31" "Q43229", X "P31" Y, Y "P31" "Q215627" => !
 
 .run-md disjointness-violations-top4-length1-2
