+++ sample_scripts/wikidata-split-order.zph
@@ -5,7 +5,7 @@
 .name "is subclass of" wikidata P279
 # .name "transitive relation" wikidata Q18647515
 
-# R is instance of Q18647515, X R Y, Y R Z => X R Z
-# I is instance of S, S is subclass of C => I is instance of C
+# R "is instance of" "Q18647515", X R Y, Y R Z => X R Z
+# I "is instance of" S, S "is subclass of" C => I "is instance of" C
 
-I is subclass of C, I is instance of C => !
+I "is subclass of" C, I "is instance of" C => !
